Moved databasePage to pages folder.  Added some missing tables
