+++ frontend/src/pages/databasePage.tsx
@@ -1,7 +1,7 @@
 ﻿import * as React from 'react';
 import {Link} from "react-router-dom";
 import {Button} from "@mui/material";
-import { supportedTables } from "../helpers/FrontendDatabase"
+import { supportedTables } from "../helpers/FrontendDatabase.tsx"
 
 const Databasepage: React.FC = () => {
